requirements.txt
pytz
django==1.7.3
jsonpickle
pillow
django-suit
django-ckeditor-updated
django-select2
django-filter
django-htmlmin
django-debreach
django-ipware
django_compressor
djangorestframework
markdown
markdown2
easy-thumbnails
python-slugify